// client/src/models/enums/ForumCategory.ts
export enum ForumCategory {
    GENERAL = 'general',
    ALGORITHMS = 'algorithms',
    INTERVIEWS = 'interviews',
    TECHNOLOGIES = 'technologies',
    CAREERS = 'careers',
    UNIVERSITY = 'university',
    FEEDBACK = 'feedback',
}

export const ForumCategoryList = Object.freeze(Object.values(ForumCategory));